Create SyntaxUtils file
